Renamed test_device_memset --> test_host_memset test_memset --> test_device_memset test_memcpy --> test_device_memcpy
