+++ regression/cuda/test_device_memcpy.h
@@ -25,8 +25,8 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef TEST_MEMCPY_H
-#define TEST_MEMCPY_H
+#ifndef TEST_DEVICE_MEMCPY_H
+#define TEST_DEVICE_MEMCPY_H
 
 
 #include <stdio.h>
