+++ regression/cuda/test_device_memset.c
@@ -25,28 +25,37 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
+/******************************************************************************/
+/* Runs the device_memset kernel a grid of 2x2 tile groups.                   */
+/* Host allcoates space on DRAM and passes the pointer and size to tiles      */
+/* Tiles fill the space with the requested value from host                    */
+/* Host the compares the values with expected.                                */
+/* Grid dimensions are prefixed at 1x1.                                       */
+/* This tests uses the software/spmd/bsg_cuda_lite_runtime/device_memset/     */
+/* manycore binary in the BSG Manycore repository.                            */
+/******************************************************************************/
+
+
 #include "test_device_memset.h"
 
 #define TEST_NAME "test_device_memset"
 #define ALLOC_NAME "default_allocator"
-#define TEST_BYTE 0xcd
-
-/*!
- * Runs a device_memset kernel on a 2x2 tile group. 
- * Device allcoates memory on device and uses hb_mc_device_memset to set to a prefixed valu.
- * Device then calls an empty kernel and loads back the meomry to compare.
- * This tests uses the software/spmd/bsg_cuda_lite_runtime/device_memset/ Manycore binary in the BSG Manycore bitbucket repository.  
-*/
+#define TEST_VALUE 0x1234
 
 
 int kernel_device_memset () {
-	bsg_pr_test_info("Running the CUDA Device Memset Kernel on a grid of one 2x2 tile group.\n\n");
+	bsg_pr_test_info("Running the CUDA Memset Kernel "
+                         "on a grid of 2x2 tile groups.\n\n");
 	int rc;
 
-	/*****************************************************************************************************************
-	* Define path to binary.
-	* Initialize device, load binary and unfreeze tiles.
-	******************************************************************************************************************/
+	srand(time); 
+
+
+
+        /**********************************************************************/
+        /* Define path to binary.                                             */
+        /* Initialize device, load binary and unfreeze tiles.                 */
+        /**********************************************************************/
 	hb_mc_device_t device;
 	rc = hb_mc_device_init(&device, TEST_NAME, 0);
 	if (rc != HB_MC_SUCCESS) { 
@@ -54,7 +63,8 @@ int kernel_device_memset () {
 		return rc;
 	}
 
-	char* elf = BSG_STRINGIFY(BSG_MANYCORE_DIR) "/software/spmd/bsg_cuda_lite_runtime" "/device_memset/main.riscv";
+	char* elf = BSG_STRINGIFY(BSG_MANYCORE_DIR) "/software/spmd/bsg_cuda_lite_runtime"
+                                                    "/device_memset/main.riscv";
 	rc = hb_mc_device_program_init(&device, elf, ALLOC_NAME, 0);
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to initialize program.\n");
@@ -62,54 +72,58 @@ int kernel_device_memset () {
 	}
 
 
-
-	/*****************************************************************************************************************
-	* Allocate memory on the device for A and set it to TEST_VAL
-	******************************************************************************************************************/
+        /**********************************************************************/
+	/* Allocate memory on the device for A_ptr.                           */
+        /**********************************************************************/
 	uint32_t N = 64;
-	eva_t A_device; 
-	rc = hb_mc_device_malloc(&device, N * sizeof(uint8_t), &A_device); /* allocate A_ptr on the device */
+
+	hb_mc_eva_t A_device; 
+	rc = hb_mc_device_malloc(&device, N * sizeof(uint32_t), &A_device);
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to allocate memory on device.\n");
 		return rc;
 	}
 
-
-	rc = hb_mc_device_memset(&device, &A_device, TEST_BYTE, N * sizeof(uint8_t));
+	// Set the entire space to zero
+	rc = hb_mc_device_memset(&device, &A_device, 0, N * sizeof(uint8_t));
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to set memory on device.\n");
 		return rc;
 	} 
 
 
-	/*****************************************************************************************************************
-	* Define block_size_x/y: amount of work for each tile group
-	* Define tg_dim_x/y: number of tiles in each tile group
-	* Calculate grid_dim_x/y: number of tile groups needed based on block_size_x/y
-	******************************************************************************************************************/
+        /**********************************************************************/
+	/* Define block_size_x/y: amount of work for each tile group          */
+	/* Define tg_dim_x/y: number of tiles in each tile group              */
+	/* Calculate grid_dim_x/y: number of                                  */
+        /* tile groups needed based on block_size_x/y                         */
+        /**********************************************************************/
 	hb_mc_dimension_t tg_dim = { .x = 2, .y = 2 }; 
 
 	hb_mc_dimension_t grid_dim = { .x = 1, .y = 1 };
 
 
-	/*****************************************************************************************************************
-	* Prepare list of input arguments for kernel.
-	******************************************************************************************************************/
-	int argv[1];
+        /**********************************************************************/
+	/* Prepare list of input arguments for kernel.                        */
+        /**********************************************************************/
+	int argv[3] = {A_device, TEST_VALUE , N};
 
-	/*****************************************************************************************************************
-	* Enquque grid of tile groups, pass in grid and tile group dimensions, kernel name, number and list of input arguments
-	******************************************************************************************************************/
-	rc = hb_mc_application_init (&device, grid_dim, tg_dim, "kernel_device_memset", 0, argv);
+
+        /**********************************************************************/
+	/* Enquque grid of tile groups, pass in grid and tile group dimensions*/
+        /* kernel name, number and list of input arguments                    */
+        /**********************************************************************/
+	rc = hb_mc_application_init (&device, grid_dim, tg_dim, "kernel_device_memset", 3, argv);
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to initialize grid.\n");
 		return rc;
 	}
 
 
-	/*****************************************************************************************************************
-	* Launch and execute all tile groups on device and wait for all to finish. 
-	******************************************************************************************************************/
+
+        /**********************************************************************/
+	/* Launch and execute all tile groups on device and wait for finish.  */ 
+        /**********************************************************************/
 	rc = hb_mc_device_tile_groups_execute(&device);
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to execute tile groups.\n");
@@ -117,24 +131,22 @@ int kernel_device_memset () {
 	}	
 
 
-
-	/*****************************************************************************************************************
-	* Copy result matrix back from device DRAM into host memory. 
-	******************************************************************************************************************/
-	uint8_t A_host[N];
+        /**********************************************************************/
+	/* Copy result vector back from device DRAM into host memory.         */
+        /**********************************************************************/
+	uint32_t A_host[N];
 	void *src = (void *) ((intptr_t) A_device);
-	void *dst = (void *) &A_host[0];
-	rc = hb_mc_device_memcpy (&device, (void *) dst, src, N * sizeof(uint8_t), HB_MC_MEMCPY_TO_HOST);
+	void *dst = (void *) &A_host;
+	rc = hb_mc_device_memcpy (&device, (void *) dst, src, N * sizeof(uint32_t), HB_MC_MEMCPY_TO_HOST);
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to copy memory from device.\n");
 		return rc;
 	}
 
 
-
-	/*****************************************************************************************************************
-	* Freeze the tiles and memory manager cleanup. 
-	******************************************************************************************************************/
+        /**********************************************************************/
+        /* Freeze the tiles and memory manager cleanup.                       */
+        /**********************************************************************/
 	rc = hb_mc_device_finish(&device); 
 	if (rc != HB_MC_SUCCESS) { 
 		bsg_pr_err("failed to de-initialize device.\n");
@@ -144,8 +156,11 @@ int kernel_device_memset () {
 
 	int mismatch = 0; 
 	for (int i = 0; i < N; i++) {
-		if (A_host[i] != TEST_BYTE) { 
-			bsg_pr_err(BSG_RED("Mismatch") ": -- A[%d] = 0x%08" PRIx32 "\t Expected: 0x%08" PRIx32 "\n", i , A_host[i], TEST_BYTE);
+		if (A_host[i] != TEST_VALUE) {
+			bsg_pr_err(BSG_RED("Mismatch: ") "A[%d] =  0x%08" PRIx32 "\t Expected: 0x%08" PRIx32 "\n",
+                                                         i,
+                                                         A_host[i],
+                                                         TEST_VALUE);
 			mismatch = 1;
 		}
 	}
