+++ regression/cuda/test_host_memset.h
@@ -25,8 +25,8 @@
 // (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
 // SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
-#ifndef TEST_MEMSET_H
-#define TEST_MEMSET_H
+#ifndef TEST_HOST_MEMSET_H
+#define TEST_HOST_MEMSET_H
 
 
 #include <stdio.h>
