+++ regression/cuda/tests.mk
@@ -49,13 +49,13 @@ UNIFIED_TESTS += test_barrier
 INDEPENDENT_TESTS += test_binary_load_buffer
 INDEPENDENT_TESTS += test_empty_parallel
 INDEPENDENT_TESTS += test_multiple_binary_load
-INDEPENDENT_TESTS += test_device_memset
+INDEPENDENT_TESTS += test_host_memset
 INDEPENDENT_TESTS += test_stack_load
 INDEPENDENT_TESTS += test_dram_load_store
 INDEPENDENT_TESTS += test_dram_host_allocated
 INDEPENDENT_TESTS += test_dram_device_allocated
-INDEPENDENT_TESTS += test_memset
-INDEPENDENT_TESTS += test_memcpy
+INDEPENDENT_TESTS += test_device_memset
+INDEPENDENT_TESTS += test_device_memcpy
 INDEPENDENT_TESTS += test_vec_add
 INDEPENDENT_TESTS += test_vec_add_parallel
 INDEPENDENT_TESTS += test_vec_add_parallel_multi_grid
